[fix] Moved receiver functions to receiver.py
