+++ openwisp_firmware_upgrader/templates/admin/firmware_upgrader/batch_upgrade_operation_change_form.html
@@ -16,8 +16,8 @@
 <!-- Upgrade Operations Section -->
 <h1 class="upgrade-operations-title">{% trans "Upgrade Operations" %}</h1>
         
-        <!-- OpenWISP Filter Section -->
-        {% if filter_specs %}
+    <!-- OpenWISP Filter Section -->
+      {% if filter_specs %}
         <div id="ow-changelist-filter">
             <div class="filters-bottom">
                 <span class="left-arrow">
